refactor: 컴포넌트 구조 변경 및 파일 구조 변경
