+++ src/App.js
@@ -1,10 +1,10 @@
 import "./App.css";
-import HomePage from "./pages/home/index";
+import Shared from "./pages/Shared";
 
 function App() {
   return (
     <div className="App">
-      <HomePage />
+      <Shared />
     </div>
   );
 }
