+++ src/pages/Shared.js
@@ -1,16 +1,16 @@
 import React, { useEffect, useState } from "react";
-import Header from "../../components/Header/Header";
-import Footer from "../../components/Footer/Footer";
-import { getSampeUser, getSampleFolder } from "../../utils/Api";
-import FolderTilte from "../../components/FolderTitle/FolderTilte";
-import CardList from "../../components/CardList/CardList";
+import Header from "../components/Header/Header";
+import Footer from "../components/Footer/Footer";
+import { getSampeUser, getSampleFolder } from "../utils/Api";
+import FolderTilte from "../components/FolderTitle/FolderTilte";
+import LinkContainer from "../components/links-container/Container";
 
-function Index() {
+function Shared() {
   // 함수 이름을 대문자로 변경
   const [user, setUser] = useState(null);
-  console.log(user);
 
   const [folder, setFolder] = useState(null);
+  console.log(folder);
 
   useEffect(() => {
     getSampeUser().then(setUser);
@@ -21,10 +21,10 @@ function Index() {
     <main>
       <Header user={user} />
       <FolderTilte folderObject={folder} />
-      <CardList folderObject={folder} />
+      <LinkContainer folderObject={folder} />
       <Footer />
     </main>
   );
 }
 
-export default Index;
+export default Shared;
